requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow
apache-airflow-providers-postgres
apache-airflow-providers-http
psycopg2-binary
SQLAlchemy
yfinance
pandas
numpy
loguru
pendulum
